+++ src/main/java/com/knocksea/see/inquiry/repository/InquiryRepository.java
@@ -4,6 +4,6 @@ import com.knocksea.see.inquiry.entity.Inquiry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface InquiryRepository extends JpaRepository<Inquiry, Integer>{
+public interface InquiryRepository extends JpaRepository<Inquiry, Long>{
 
 }
